Skip SuspendCheckHandler with interpreter
Implicit suspend check is handled now in an exception handler, with
reading the faulty instruction opcode to decide if code needs to
suspend or it is a normal fault.
With execute-only mode coming to Android, all (non-generated) code
sections will lose read permission so reading the instruction opcode
will generate a fault. This is still fine with JIT/AOT as those
generated instructions won't be execute only (ART does tricks like this
one, so we would most likely loose performance while making generated
code execute-only compatible), but causes a crash if there is a null
pointer exception in Nterp, as the Nterp interpreter code is part of
libart.
Since interpreter does not use implicit suspend checks, we can just
skip SuspendCheck handler if the fault PC is in Nterp..

Bug: 379071663
Test: art/test/testrunner/testrunner.py --target --64 --interpreter --jit
--optimizing
Flag: EXEMPT BUGFIX
      This should be a no-op and avoids a null pointer exception
      when execute-only memory is enabled.
Change-Id: Ia438cc82c2f099a5e5aa67c19f0696bc384324ca

Change-Id: If7e7e4e0564442d2fef3aae0ae29fc0bd352463e