Add Art BCP profile to fragment configurations.
Previously, boot image profiles were imported using a ad-hoc array. This CL changes it so that a module can declare its profile information using a path instead.

This improves the correctness of boot image profiles by ensuring that the profile is declared directly in the definition of the module that provides the profile and that it is only added if the profile is present.

Bug: 342163020
Test: m
Test: atest ArtServiceTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I121699fc2b4ce7b487602ad7699bf4d9aa3f31a4