Consider priority boosting in suspension wait
We use a longer thread suspend timeout in WaitForSuspendBarrier when
the requestor has low priority. But ScopedPriorityChange now commonly
raises the priority around SuspendAll calls, which causes us to ignore
the initial low priority. Save the original priority in the Thread
structure, and check it first in WaitForSuspendBarrier.

This should be roughly performance neutral on average, since it
sometimes saves us a getpriority system call, in return for a bit of
extra bookkeeping. Due to the placement of the new 8-bit field, it
should not affect the size of the Thread structure.

Add a test for priority boosting.

Remove redundant and slightly confusing ScopedObjectAccess in
Runtime::Init.

Test: atest art_standalone_runtime_tests on cuttlefish
Flag: EXEMPT BUGFIX
Bug: 441175965
Change-Id: I8e0fcf3966c60cf1c0108257dd67159764738d63