Don't timeout when waiting for DumpCheckpoint unless we are aborting
For dumping the stack traces of all threads, we install a DumpCheckpoint
and wait for all threads to run the checkpoint within the specified
timeout. The DumpCheckpoint is a stack allocated object and it is
released when we return from waiting from the checkpoints. If there was
a timeout and the threads haven't finished running a checkpoint this
could cause memory corruption issues.

Since we use this path to dump stacks when aborting, we do want a
timeout so we get some stack traces before aborting. So, this CL fixes
it by:
1. Not having a timeout in regular builds if we are not aborting
2. Have a timeout if we are aborting or in debug builds. In both cases
   we abort on a timeout.
3. Allocate the DumpCheckpoint on heap and delete it only if we
   successfully ran the checkpoint on all threads without a timeout.

This could leak the DumpCheckpoint object when there is a timeout, but
we are aborting there anyway and it is better to leak it than have
memory corruption issues.

Bug: 438339018
Test: art/test.py
Flag: EXEMPT BUGFIX
Change-Id: I17b892fe6d6d6fce926a2c4fa3d7b4dba1b79ff3