Support tracing all methods with low-overhead tracing
Low-overhead tracing was initially designed to get traces from user
devices. Though this can be a useful tool for developers too to get fast
method tracing even though it is imprecise. This CL extends the
low-overhead method tracing to support tracing all methods and flushing
them to file. We use the regular method tracing framework to flush all
the methods. The initial low-overhead tracing was expecting traces to be
small and were maintained in-memory until they are either requested to
be flushed or we discard them since they are not needed. This doesn't
work for the full method tracing.

Bug: 441450730
Test: art/test.py
Flag: com.android.art.rw.flags.enable_profile_code_rw
Change-Id: I4c3d735bf89c2db83f93806bca8a36ec277f20d2