Be more lenient when a native JNI caller may be in an app.
In the special case where we see a platform caller on the managed stack
and a non-platform caller on the native stack, it's not enough to check
that the native caller is an app that is allowed access. It may also be
an app that's using a currently working method to circumvent hidden API
checks, and in that case we should keep it working if app native JNI
caller identification (com.android.art.flags.hiddenapi_jni_api_callers)
is disabled.

Bug: 443069921
Bug: 377676642
Test: v2/android-app-compat-engprod/csuite/top_100_app_crawl
Test: v2/android-app-compat-engprod/csuite/top_100_app_crawl_gce
Flag: com.android.art.flags.hiddenapi_platform_enforcement
Change-Id: I2bcebc56a0bdcbce5bb8bef31dfd21eb7569a72a