Allocate an extra word for chunk-info-vec...
...to avoid checking for out-of-bound access while updating the vector,
which is done in the marking phase hot path. This also simplifies a bit
code in PrepareForCompaction() when doing exclusive_scan(). But to avoid
the mis-alignment for all the arrays that we carve out of info_vec_,
moving chunk-info-vec to the end. This way we will only pay for an extra
virtual page but still have everything perfectly aligned in cache-lines.

Flag: EXEMPT PURE_REFACTOR
Test: art/test/testrunner/testrunner.py --host --debug
Change-Id: Id686f05865ddb6a9eb72fc446117608ea70c42f3