Test membarrier registration APIs in isolated process
The membarrier registration mechanism applies to the calling process and
cannot be cleared. This means the tests for this are reliant on the
order in which tests are executed and the isolate mode used by the test
harness. For example, running the art_libartbase_tests suite with these
flags:

art_libartbase_tests --no_isolate --gtest_repeat=2

produces a failure on the second iteration of
membarrier.private_expedited as the process is registered by the first
iteration.

This introduces a small helper to run logic in a child process and uses
that to isolate the tests for the membarrier registration mechanism.

Test: art_libartbase_tests
Flag: EXEMPT TEST_ONLY
Change-Id: Ia67a30393b5d4dde6d52e1ed99b9c1458ba594d4