Convert art-host-tests zip to soong
This is the final step in migrating the art-host-tests package from
Make to Soong.

The temporary test module `art-host-tests-bp` has been verified to work
correctly. This change renames it to the final `art-host-tests` name,
making it the canonical build target.

This change is paired with the removal of the legacy
build/core/make/tasks/art-host-test.mk in the build/make repository to
complete the migration.

Flag: EXEMPT PURE_REFACTOR

Bug: b/431634358
Test: m -j art-host-tests
Change-Id: I74c365f78c1559e895021373a5fb5a87a4e819a8