Convert art-host-tests zip to soong
This is the final step in migrating the art-host-tests package from
Make to Soong.

The temporary test module `art-host-tests-bp` has been verified to work
correctly. This change renames it to the final `art-host-tests` name,
making it the canonical build target.

This change is paired with the removal of the legacy
build/core/make/tasks/art-host-test.mk in the build/make repository to
complete the migration.

Flag: EXEMPT PURE_REFACTOR

Bug: b/431634358
Test: m -j art-host-tests
Change-Id: I637657695ed16c252257461879b0e54e7d652a87