Check the existence of dex files in CLCs of secondary dex files.
Before this change, when dexopting a secondary dex file, if one of the
dex files in the CLC is deleted, dexopt will fail when using
OatFileAssistant to check dexoptNeeded.

After this change, a CLC containing non-existing dex files is treated as
UNSUPPORTED_CLASS_LOADER_CONTEXT.

This change avoids the red herring in logs and metrics.

Bug: 343901555
Bug: 447314043
Test: atest ArtServiceTests
Test: atest art_standalone_artd_tests
Flag: EXEMPT BUGFIX
Change-Id: I1d222c05b50f1473f499bd946993badb67fc1116