Remove listeners only when there is an overflow
There was a bug and we removed listeners even when the buffer was not
full. That meant we ended tracing earlier than necessary in
non-streaming case.

Also fixed a bug where we were passing uint16_t for storing a 24-bit
value which was uncovered when adding a test for this (ag/36025507)

Bug: 441450730
Test: art/test.py
Flag: EXEMPT BUGFIX
Change-Id: Iefc8450b5767077223855c4b6dae619270005387