Fix concurrent_start_bytes_ for blocking gc
Set concurrent_start_bytes_ so that we do not rely on the number of
bytes allocated during a blocking GC, which is generally misleading.

This introduces running_collection_delayed_allocation_, and adjusts the
computation of running_collection_is_blocking_, which I do not believe
was previously correct. The difference is that the latter includes
waiting for the GC for reasons other than the inability to allocate.
We can wait for a GC in many, mostly obscure, cases. The change to
running_collection_is_blocking_ affects statistics we generate about
blocking GCs.

Change numerous GUARDED_BY and REQUIRES annotations to mention
gc_complete_lock instead of *gc_complete_lock_ for consistency. I
believe it does not matter, but I otherwise had no idea which to use.
The documentation is unclear, but kind of, sort of, weakly suggests
using the pointer directly when convenient.

Flag: EXEMPT BUGFIX
Bug: 443342017
Test: Build and boot, host run-tests.
Change-Id: Id5933cc18e56d4390404872aa83660b90484254c