Assign the source name object when emitting dex pc
When there is no debug info the line number of a StackTraceElement is
set to the dex_pc and the source name object of the StackTraceElement is
set to null.

This change instead sets the source name object to the source file of
the declaring class when the dex_pc is used.

Motivation:

R8 and other optimizers encode the id of the mapping file in the source
file of each class. This makes it possible to identify the mapping file
that is required for retracing an obfuscated stack trace from only
looking at the obfuscated stack trace.

However, since ART currently does not emit the source file for methods
that do not have any debug info, R8 must emit debug info for all methods
to ensure that the id of the mapping file can be found from stack
traces.

By emitting the source file when the dex_pc is used it is possible for
R8 to drop all debug info when the min API is high enough, which leads
to smaller DEX.

Bug: 146565491
Flag: EXEMPT BUGFIX
Test: existing
Change-Id: Ib135bbfbb5aad2958ab63248dbb10b973acbe4ea