Optimizing: Clean up use search in `LiveInterval`.
The `FirstRegisterUseAfter()` should not do the check for
required register at definition as that has nothing to do
with actual uses and most callers don't need it anyway.
Refactor that check out of the function.

Remove the defining position check from `FirstUseAfter()`
as it's always false for both callers.

Test: testrunner.py -b --host-gtest --host --optimizing
Flag: EXEMPT PURE_REFACTOR
Change-Id: I788f452dda8e2014d4b1967b33246b7e71475309