Optimizing: Fix `RequiresRegisterForDefinitionAt()`.
Fix a latent bug that could trigger a `DCHECK()` but seems
unreachable for the arguments that the current register
allocator passes to this function.

Unify the results across all physical register types. This
was previously returning true for fixed core registers but
false for fixed FP registers.

Test: testrunner.py -b --host-gtest --host --optimizing
Flag: EXEMPT BUGFIX
Change-Id: Id7e7a98e6cc8d75a2758550e15ca4e96e4e55349