ART: Clean up `check-cast`/`instance-of` verification.
Remove a misinterpreted check for conflict (meaning `void`)
and pull anoher check earlier to cover that case. Change the
failure message to use `DexFile::PrettyType()` and pull it
out to a `NO_INLINE` helper function.

Note that we could do this check in an earlier stage of the
verification but that would need a new verification flag.
We're currently using 31 bits of the `uint32_t` flags and
avoiding this simple check on revisiting one of these two
instructions would not speed up the verification enough to
be worth using the last available bit.

Test: testrunner.py -b --host-gtest --host --optimizing
Flag: EXEMPT PURE_REFACTOR
Change-Id: Id950e27ac0da4b2d457a95bd5088ab6b1b86ca0a