Add a missing GUARDED_BY annotation for buf_ in TraceWriter
buf_ is used to hold encoded data for trace events. TraceWriter thread
uses it when encoding trace events. When stopping the trace we also
use this on the thread that is stopping. At that point everything is
suspended and TraceWriter thread exited as well. So it was always used
safely but adding an annotation to guard against any future
modifications.

Test: art/test.py -t 2246
Flag: EXEMPT BUGFIX
Change-Id: I734f164607e5c78b46c574a85c070ab97ffe777e