x86_64: use SystemArrayCopyOptimizations for primitive arrays.
If we know at compile time that source and destination are the same
array, we should not generate intrinsic at all, as the first thing the
generated code does is an unconditional jump to the slow path. If we
know that one of them or both are non-null, we should drop null checks.

Test: see the changes in 537-checker-arraycopy
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b \
        -t 537-checker-arraycopy

Bug: 388034836
Flag: EXEMPT BUGFIX
Change-Id: I6f356eca79d6a09eb881ed2af24445a6f172177d