Optimize profile saver (step 2).
1. Don't collect hot methods from JitCodeCache. We are already
   collecting hot methods from ArtMethod::PreviouslyWarm.
2. Add suspension checks to every profile info visit, to avoid holding
   `Locks::mutator_lock_` for too long.

For #1 to work, this CL also changed where we mark a method as
previously warm, to make sure that the flag reflects whether a method
has ever been baseline compiled.

Before this change, the flag is set by nterp and missing in two cases:

1. When a method is promoted from the switch interpreter to baseline.
2. When a method is promoted from nterp to baseline and the method is in
   shared memory.

After this change, the flag is set by JIT whenever a method is enqueued
for baseline compilation.

Bug: 453630216
Test: art/test/testrunner/testrunner.py --target --64 -t \
  595-profile-saving (in chroot)
Test: art/test/testrunner/testrunner.py --target --64 -t \
  2271-profile-inline-cache (in chroot)
Test: art/tools/run-gtests.sh -j8 (in chroot)
Flag: EXEMPT BUGFIX

Change-Id: I3e6214b026f5ccd081e2fe00fa0a20ccd9c133cc