Revert^2 "Move MH callsite check from intrinsics to IR."
This reverts commit 3c91d4651c1e83077faeae30991fcff301370b2d.

Reason for revert: fixed --no-image tests.

Another way to solve that was to hack artQuickResolutionTrampoline
in such a way that when it is called for an invoke-polymorphic
method to return Invokers::checkExactType method, but keeping
type check in the intrinsics for no boot image cases seems to be
cleaner.

Bug: 297147201
Bug: 454527749

Flag: EXEMPT PURE_REFACTOR
Test: ./art/test/testrunner/testrunner.py -b --target --64
Test: ./art/test/testrunner/testrunner.py -b --host --64
Test: ./art/test/testrunner/testrunner.py -b --host --jit --64
Test: ./art/test/testrunner/testrunner.py -b --target --no-image --64
Test: ./art/test/testrunner/testrunner.py -b --host --no-image --64

Change-Id: I1c87251a598f3e4793f6f1a2040b9956de3f6ba7