x86_64: relax checker expression in 537-checker-arraycopy.
Recent commit 2f1f88417e98d7e16f6e9c2dddeffdf9565bd274 has broken test
537-checker-arraycopy when running with `--no-image` because the checker
expression was too restrictive and didn't match a slight variation in
the expected code.

This change relaxes checker expression. The new expression no longer
captures the relation between register in MOV and CALL instructions, but
capturing it requires a more complex expression, and it's not the point
of the test (the goal is to check that the intrinsic is not generated).

Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b \
        -t 537-checker-arraycopy
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b \
        -t 537-checker-arraycopy --no-image

Bug: 388034836
Flag: EXEMPT TEST_ONLY
Change-Id: Iba8e23594271849a14d50e75be8f2bd591d50544