x86: use SystemArrayCopyOptimizations for primitive arrays.
If we know at compile time that source and destination are the same
array, we should not generate intrinsic at all, as the first thing the
generated code does is an unconditional jump to the slow path. If we
know that one of them or both are non-null, we should drop null checks.

Test: art/test/testrunner/testrunner.py --host --32 --optimizing -b \
        -t 537-checker-arraycopy
Test: art/test/testrunner/testrunner.py --host --32 --optimizing -b \
        -t 537-checker-arraycopy --no-image
Test: also checked that the above tests still pass with `--64`

Test: art/test/testrunner/testrunner.py --host --32 --optimizing -b \
        -t 641-checker-arraycopy

Bug: 388034836
Flag: EXEMPT BUGFIX
Change-Id: If15114f4546bb63dd4cfb736bc9b6fb50d23e830