DCHECK that IsPowerOfTwo is not called with 0
The current implementation assumes that IsPowerOfTwo will not be
called with 0, and will return the opposite value (i.e true instead
of false).

There's another implementation in art/sigchainlib/sigchain.cc which
only has one use in a static assert, and it is isolated from the
one this CL updates.

Bug: 448020537
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Flag: EXEMPT BUGFIX
Change-Id: Ib50e27ea8e31fa37d0fb161997d5b38817ace313