Remove unnecessary ScopedGCCriticalSection
Expanding slightly on prior code comments, I believe these were
addressing the following scenario:

1) GC thread is running a checkpoint to disable weak reference access.
   It disabled weak reference access for (2).
2) A thread is holding the dex lock and waiting for weak reference
   access. Which means it's waiting for (1).
3) Another thread is runnable, and waiting for the dex lock, which means
   it's waiting for (2) to release the lock.

This means that (1) is waiting for (3) to notice the suspension request,
Thus (3) is waiting for (2), which is waiting for (1), which is waiting
for (3), and we were stuck.

Although this used to be possible, aosp/1798377 replaced the checkpoint
to disable weak reference access with a stop-the-world pause to address
other similar deadlocks. As explained in the "Waiting" section of
mutator_gc_coord.md, this is now impossible. In particular, Thread (2)
could not be interrupted while it is holding the dex lock, and is about
to wait for weak reference access. To do so it would have had to release
the mutator lock, which would have violated lock ordering.

As pointed out by the bugs below, the ScopedGCCriticalSections introduced
significant jank.

This effectively reverts aosp/692442.

Flag: EXEMPT BUGFIX
Bug: 452684724
Bug: 399287133
Test: Treehugger, some manual testing by partners
Change-Id: I73dc50a48792c12b19f3811475e5dc648385d74d