arm64/nterp: Faster opcode dispatch.
Create 4 distinct sets of opcode handlers starting at 4KiB
aligned addresses and 20KiB apart, so that one of them is at
a 16KiB aligned address. Form handler address by inserting
the opcode bits to the bits 13:6 of the `xIBASE` register
pre-initialized to point to the 16KiB aligned handler set
(except for the replacable bits).

Since slow paths were just a bit over 4KiB, implement a way
to add a slow path shared across different nterp prefixes
and use it for part of the `instance-of` slow path.

Benchmarking with the following code added to `001-Main`,
we see almost 12% reduction in the reported time.

    public static void main(String args[]) {
        timeArraySum(new int[10000]);
        timeArraySum(new int[100000]);
        timeArraySum(new int[1000000]);
        timeArraySum(new int[10000000]);
        timeArraySum(new int[50000000]);
    }

    public static int timeArraySum(int[] a) {
        long start = System.nanoTime();
        int sum = arraySum(a);
        long end = System.nanoTime();
        System.out.println("length=" + a.length + " time=" + (end - start));
        return sum;
    }

    public static int arraySum(int[] a) {
        int sum = 0;
        for (int v : a) {
            sum += v;
        }
        return sum;
    }

Test: testrunner.py --target --64 --interpreter
Test: Repeat with `ART_HEAP_POISONING=true`.
Flag: EXEMPT PURE_REFACTOR
Change-Id: I1ee990d7d69d892aca9901032117a2ad8bcc5349