Speed up ClassTable::Resize()
... By caching the class descriptor hash.

Do this by reusing Class::cl_init_thread_id_ to avoid growing the class
structure.

We only cache hash codes that do not have all of the leading 10 bits
set. Thus, by storing the bit-wise complement, we can distinguish cached
hash values from 22-bit thread ids.

Note that the hash function maps very short names to small hash values,
ensuring that minified class names are basically certain to be
cacheable.

We always cache the hash value after a class is initialzed, ensuring
that this operation doesn't dirty the class object later.

Empirically, we get > 99% hits on the cached descriptor hashes in
Class::DescriptorHash(), even during boot, as measured by temporarily
added log entries in and after the first if-statement.

We narrow the interval in which ClInitThreadId is set for inheritance
cycle detection. This significantly reduces interfence with the use as
a hash code. It also technically removes the need for clearing the
field when writing out the class. But we want to overwite it with the
hash code anyway to minimize later modification of the class object.

Flag: EXEMPT BUGFIX
Bug: 454767701
Bug: 458973673
Test: Build and boot cuttlefish, Treehugger
Test: Manually measure boot hit rate for cached hash values
Change-Id: I83dc50f1a6c2dc4e45c859604b7d764dbf512c51