Veridex may ignore aconfig guarded blocks
Use --ignore-aconfig-guards option to ignore aconfig guarded blocks
from the hidden api inspection. With the option, veridex detects the
fields or methods for aconfig flags from arguments of instructions.
Any methods that use the aconfig flags are also marked as aconfig
methods. The aconfig guard finder looks for the `if` blocks with the
aconfig arguments and skips those blocks.

Bug: 456385928
Test: run veridex on apps with aconfig flagged APIs.
      compare the result with and without the --ignore-aconfig-guards
Flag: EXEMPT HOST_MODE
Change-Id: I9a15ae51dcf2ff41d3dcb3d3552b49727e91c94d