Special case catch blocks for loop analysis
Potentially all blocks in a try can flow into the corresponding
catch blocks. Loop analysis was not using this, which meant that
a block was not considered to be part of a loop (which in turn
caused other issues).

Also, the block that starts the try shouldn't be considered to be
part of the catch block.

Test: Compile the APK in the bug
Flag: EXEMPT BUGFIX
Bug: 453364730
Change-Id: I513666ec6d4a37c15e379ec825aa623627f98d82