Synchronize Virtual Threads
A thin lock id is allocated for every virtual thread.

If a virtual thread owns a thin lock, and is mounted on a platform thread,
the platform thread needs to be suspended for lock inflation.

During lock inflation, we make sure that the virtual thread stays
unmounted or mounted on a platform thread.

Flag: com.android.art.flags.virtual_thread_impl_v1
Bug: 454244174
Test: testrunner.py
Change-Id: Ieab9f532cd44e86e9ebfc99ca94bcc6f4752ec1a