arm: SystemArrayCopy: allow backward copy in tha same array.
Previously SystemArrayCopy for primitive types would fallback to the
libcore implementation if source and destination were the same array.
This change relaxes this restriction and allows backward copy. Note
that forward copy for overlapping arrays is problematic with the current
intrinsic implementation, as it would overwrite elements that haven't
been copied yet.

This change also removes code duplication between the checks in the
primitive and generic versions, as well as code duplication in creating
locations for primitive and generic versions.

Also, this fixes a bug in the primitive version: unintentional BNE
instead of BEQ in the null checks.

Test: added more test cases and updated 537-checker-arraycopy.
Test: art/test.py -r --target --optimizing

Flag: EXEMPT BUGFIX
Bug: b/388034836
Change-Id: I01d4e1d178a056253d4a254a6f86ed829d4f7208