Revert^2 "Special case catch blocks for loop analysis"
This reverts commit 472e69df41e9ad8b693f03d889d203dc1aa8e9ed.

Reason for revert:

Fix for 463606233 is to run ComputeTryBlockInformation before
AnalyzeLoops as now AnalyzeLoops uses that information.

Fix for 453364730 is to prioritize non-backedges before backedges.
Tests 594- and 596- tests updated to not check a specific linear
order for linearization, as linearization is not guaranteed to
produce a valid linear order for graphs with irreducible loops.

Bug: 463527925
Bug: 453364730
Test: Compile the APK in bug 463606233
Change-Id: Ic86ae213959ba2142d2b51c13cf2ab880327265f