arm64: SystemArrayCopy: allow backward copy in tha same array.
Previously SystemArrayCopy for primitive types would fallback to the
libcore implementation if source and destination were the same array.
This change relaxes this restriction and allows backward copy. Note
that forward copy for overlapping arrays is problematic with the current
intrinsic implementation, as it would overwrite elements that haven't
been copied yet.

This change also removes code duplication between the checks in the
primitive and generic versions, as well as code duplication in creating
locations for primitive and generic versions.

Test: updated 537-checker-arraycopy.
Test: art/test.py -r --target --optimizing --64

Flag: EXEMPT BUGFIX
Bug: b/388034836
Change-Id: I11e7c22fd9b814762f5303514e563cd455957230