Fix skipping method exit callbacks for top frame when required
We need to skip method exit callbacks if we are deopting after method
exit callback is called. This can happen when method exit callback sets
a breakpoint on the caller or pops the frame etc., We used to check for
frame depth being 0 to skip method exit callbacks. We may have a runtime
frame at depth 0 and hence we were missing the actual frame. Changed the
check to be more robust by not depending on the depth.

Bug: 455851515
Test: art/test.py -t 2399
FLAG: EXEMPT BUGFIX
Change-Id: I4083bb5e8a9d71897d203b0f6e039457aaa8e4c8