arm/arm64: Do not fold shifts by 0 to shifter op.
On ARM, shifter operand cannot encode ASR #0 or LSR #0, the
immediate 0 represents ASR #32 or LSR #32. Note that LSL #0
is valid but we still avoid creating the shifter operand
instruction for `<< 0` as we expect this to be optimized
away by the instruction simplifier, except in rare cases.

We keep using LSL #0 for type conversion to long.

Test: m test-art-host-gtest
Test: Run `dex2oat` on the APK from the bug.
Bug: 452626599
Flag: EXEMPT BUGFIX
Change-Id: Id831f3ce583d90783b1d2827391194b6a37bc4c0