riscv64: SystemArrayCopy: allow backward copy in tha same array.
Previously SystemArrayCopy for primitive types would fallback to the
libcore implementation if source and destination were the same array.
This change relaxes this restriction and allows backward copy. Note
that forward copy for overlapping arrays is problematic with the current
intrinsic implementation, as it would overwrite elements that haven't
been copied yet.

This change also removes code duplication between the checks in the
primitive and generic versions, as well as code duplication in creating
locations for primitive and generic versions.

Change temporary register use in the code extracted into the helper
`CheckSystemArrayCopyNullOrOverlap`: reuse the same temporary for
loading `src_pos_constant` and for `EmitBlt32` (previously different
ones were used), and use a different temporary when called from
`SystemArrayCopyPrimitive`.

Test: updated 537-checker-arraycopy.
Test: art/test.py -r --target --optimizing --64 011 537 610 641 646 138
      # using on-VM approach desribed in art/test/README.chroot_vm.md
      # tests 011 537 610 641 646 138 are most relevant for arraycopy

Flag: EXEMPT BUGFIX
Bug: b/388034836
Change-Id: Ie53b965e7613a6727b37351fcf5ab88403997a5b