Fix RTP for invoke-polymorphic
Treat InvokePolymorphic separately. We cannot use the resolved
method as it may be incorrect.

When there's an explicit type check, the invoke itself will
be treated as it returns j.l.Object and the type check will
type the result.

In the rest of the cases, the HInvokePolymorphic will do the
type checking so we can assume that the return value is the
one from the proto.

Bug: 464219295
Fixes: 464219295
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Flag: EXEMPT BUGFIX
Change-Id: I5283219eba16c2d8df82623e091e3591183fbb9f