Call ObjectFree callbacks before allowing access to ObjectTagTable
If we allow access to ObjectTagTable before ObjectFree callbacks then
the user might see that an object has been collected before object free
callback was called. It is safe to change the order. ObjectFree
callbacks aren't expected use tag related API, so it is okay to not
allow access to the tag table. Only limited functionality like jvmti
env related memory management and raw monitors are allowed in these
callbacks.

This CL also updates the test to be more descriptive about why the test
is failing.

Bug: 462512951
Test: art/test.py -t 1974 --continuous-gc
FLAG: EXEMPT BUGFIX
Change-Id: Id577553a71c166ae1c453d156a1092af2f9e1ba2