arm64: fix `VarHandleGetAndSet` for FP-type with zero argument
The path when the argument is known to be zero (0.0f or 0.0) uses fewer
temporary registers, as it does not require scratch register for the new
value (it's in the zero register XZR/WZR). Previously we forgot to use
the vacant scratch register for the old value in the zero-argument case,
and tried to use the FP register for return value instead. This resulted
in an abort due to register bank mismatch.

Test: added new test case in 712-varhandle-invocations
Test: art/test.py -r --target --64 --optimizing

Bug: 464420849
Flag: EXEMPT BUGFIX
Change-Id: I830495062e10ecc4717883df28349eb8ff39864f