ART: Unify boot and app image relocation.
Refactor `RelocateBootImage()` and `RelocateAppImage()` into
a single function. This adds
  - timing logging to boot image relocation,
  - `VarHandle` and reflection object support to app image.
Some source address `DCHECK()`s are weakened.

Primary boot image relocation may be a bit slower because we
do not specialize it with simpler visitors anymore.

Note that the code pointer relocation is rewritten in a way
that can support app images with `--multi-image`. This is
something we were considering but did not implement yet.

Test: testrunner.py -b --host-gtest --host --relocate \
      --optimizing --speed-profile
Bug: 426660175
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ia086f186f275e7db882cb7735977b3373b20df7d