Remove a incorrect DCHECK
There was a DCHECK that skip_unwind_callbacks is only set to true for
the top frame. Though this check was wrong. It was at a wrong place,
inside an if condition where skip_unwind_callbacks is already false. The
check itself is wrong, the depth should be 1 and not 2. We count frames
from 0. This check is less useful and fails if something changes in the
stack walking logic, so removing it entirely.

Bug: 455851515
Test: art/test.py
Flag: EXEMPT BUGFIX

Change-Id: I931e4b18bff0425598052dd08fec67c7f45df735