Fetch the information about method when we first see the method
When recording method traces in non-streaming mode, the information
about methods has to be written in one block in the file. We used to
record ArtMethod* that we have seen and dump the information about the
methods at the end of the trace. Since we don't ensure that ArtMethod*
will not be deallocated we might end up with an invalid pointer. To
avoid this, maintain another map that maps a method id to the method
information and collect the information when we first see the method.

Bug: 452979583, 467468063
Test: art/test.py -t 2246
Flag: EXEMPT BUGFIX
Change-Id: Iea2c68371a7c2fa16b80c80d99c828a51faf5951