Avoid Speed compilation for apps used by other apps without a cloud/prebuilt profile
When an app it used by other apps, and there is no install/prebuilt
profiles:

Before this change: Compile the shared app in `Speed` compilation mode
which takes a long time and increases the Oat/Odex size

After this change:
- The shared app will be compiled in Speed-profile mode using the local profile, while keeping he .odex and .art files private to the app
- If the shared app's APk is public, then the Vdex will be public as well, so other apps will run in Verify mode instead of run-from-apk mode which is acceptable

Bug: 431777119
Bug: 467355146
Bug: 467510491
Bug: 436095314
bug: 413284995
Test: atest ArtServiceTests
Test: Installing two apps and check the loaded files that vdex is used publicly and .odex and .art are used privately https://paste.googleplex.com/6162189960478720
Flag: EXEMPT BUGFIX
Change-Id: Ieafba3f938907c25655f83cf26892616ce6af396

----
Make vdex files world-readable if the dex file is.

The vdex file does not contain profile information, so its permissions can be relaxed to match the dex file's world-readability. This allows other apps to load and verify the app even if other artifacts are not publicly accessible.

Change-Id: Ieafba3f938907c25655f83cf26892616ce6af396