Fix boot image profile issue in pre-reboot dexopt.
There was an issue in the boot image profile generation in pre-reboot
dexopt. In dexopt_chroot_setup, we bind-mount the old /system/etc into
chroot, on top of the new /system in chroot, in order to match the old
/system/lib{,64}, which are also bind-mounted into chroot. This has a
side-effect that an old version of the boot image profile in
/system/etc is also bind-mounted into chroot, causing a mismatch with
the new version of framework jars.

This CL fixes the issue by passing a clean view of the new /system to
odrefresh to make sure odrefresh gets the new boot image profile and
other files in /system/etc.

Bug: 467980872
Test: -
  1. Flash CP1A.251125.001
  2. Build the ART module, install it, and reboot.
  3. OTA to ToT next on git_main
  4. Trigger pre-reboot dexopt and wait for it to finish.
  5. Reboot.
  6. Run `adb shell pm art dump` and see `reason=ab-ota` on many apps.
Flag: EXEMPT BUGFIX
Change-Id: I47eded1a780cb353424cb6572f891f7f4ddad9ce