Don't expect Executable::GetArtMethod to always return a non-null value
For Methods and Constructor objects that aren't constructed via the
Constructor::serializationCopy the ArtMethod* would be initialized to a
non-null value. For Constructors constructed using serializationCopy the
ArtMethod* would be set to nullptr. So handle the cases where it can be
null. Also add a test for this.

Bug: 468172541
Test: art/test.py -t 2401
Flag: EXEMPT BUGFIX
Change-Id: I261e864a15cef404a1810dbf4973c5d6086216e5