Avoid calling MoveBlackDensePageForUpdate with COPY ioctl
MoveBlackDensePageForUpdate() is meant to handle corner cases with MOVE
ioctl. It's not required with COPY. The actual fix for MOVE ioctl case
is still required. With this CL the LUCI failures are expected to go away
until the kernel on bots there get updated for MOVE ioctl to be used.

Bug: 462516380
Flag: EXEMPT BUGFIX
Test: art/test/testrunner/testrunner.py --host --debug --continuous-gc
-t 160-read-barrier-stress

Change-Id: I5daad384360b9ff369f85884d9982b1417f0fb78